IGNITE-15234 Implement ByteBuf-based message packer and unpacker (#434)
* Refactor `ClientMessagePacker` and `ClientMessageUnpacker` to work with Netty `BytBuf` directly, without any wrappers.
* Remove inheritance from library implementation.
* Remove unnecessary `ByteBuffer` allocations.

**Benchmarks Before:**

```
Benchmark                                 Mode  Cnt        Score        Error  Units
ClientMessagePackerBenchmark.packInt     thrpt    3  2868266.504 ± 173548.310  ops/s
ClientMessagePackerBenchmark.packString  thrpt    3  2663375.726 ± 190027.949  ops/s
ClientMessagePackerBenchmark.packUuid    thrpt    3  2816744.141 ± 281943.169  ops/s
```

**Benchmarks After:**

```
Benchmark                                 Mode  Cnt          Score          Error  Units
ClientMessagePackerBenchmark.packInt     thrpt    3  389568322.545 ± 48247870.394  ops/s   (135x)
ClientMessagePackerBenchmark.packString  thrpt    3   30113891.585 ±  1781762.180  ops/s   (11x)
ClientMessagePackerBenchmark.packUuid    thrpt    3  331495906.854 ± 12134046.484  ops/s   (117x)
```

NOTE: Those benchmarks include `ClientMessagePacker` constructor call.

+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/runner/app/PlatformTestNodeRunner.java
@@ -99,7 +99,7 @@ public class PlatformTestNodeRunner {
         var valCol = "val";
 
         TableDefinition schTbl = SchemaBuilders.tableBuilder(SCHEMA_NAME, TABLE_NAME).columns(
-                SchemaBuilders.column(keyCol, ColumnType.INT32).asNonNull().build(),
+                SchemaBuilders.column(keyCol, ColumnType.INT64).asNonNull().build(),
                 SchemaBuilders.column(valCol, ColumnType.string()).asNullable().build()
         ).withPrimaryKey(keyCol).build();
